Move printing of sequences into ui module
